==> 7_Identifying_Commons.Rmd <==
---
title: "Examine Benzene RTs"
author: "Wesley Burr"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This file requires that the 1_Extract_Setup.Rmd and 1_UN_Extract_Setup.Rmd
files have been run. This sets up the rda files in dat and dat_un so this
can work. 

## Extract the RTs for the Max SN Bromobenzene (Aligned)

```{r}
library("tidyverse")
bromo_Morgue <- vector("list", length = 9)
for(k in c(1:4, 6:9)) {
  load(paste0("./dat/M_DONOR_", k, ".rda"))
  M_DONOR_ <- get(x = paste0("M_DONOR_", k))
  bromo <- M_DONOR_ %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(QuantSN == max(QuantSN))
  bromo_Morgue[[k]] <- bromo 
}
names(bromo_Morgue) <- paste0("DONOR_", 1:9)
```


```{r}
bromo_REST <- vector("list", length = 9)
for(k in c(1:4, 6:9)) {
  load(paste0("./dat/RH", k, ".rda"))
  RH_ <- get(x = paste0("RH", k))
  bromo <- RH_ %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(QuantSN == max(QuantSN))
  bromo_REST[[k]] <- bromo
}
names(bromo_REST) <- paste0("RH_", 1:9)
```

## Cleanup, Make Excel

```{r}
s1 <- do.call("rbind", bromo_Morgue)
s1 <- s1 %>% select(Sample, RT1, RT2)
s2 <- do.call("rbind", bromo_REST)
s2 <- s2 %>% select(Sample, RT1, RT2)
bromo <- rbind(s1, s2)
bromo <- t(bromo)
bromo <- data.frame(bromo)
names(bromo) <- bromo[1, ]
bromo <- bromo[-1, ]
bromo <- apply(bromo, MAR = 2, FUN = as.numeric)
bromo <- as.data.frame(bromo)
bromo
```

## Write Out File

```{r}
write.csv(file = "Checking_Bromobenze_IS.csv", bromo, row.names = FALSE)
```


```{r}
plot(x = unlist(bromo[1, ]),
     y = unlist(bromo[2, ]),
     xlab = "RT1", 
     ylab = "RT2")
```

## Extract the RTs for the Max SN Bromobenzene (Unaligned)

```{r}
bromo_Morgue <- vector("list", length = 9)
for(k in c(1:4, 6:9)) {
  load(paste0("./dat_un/M_DONOR_", k, ".rda"))
  M_DONOR_ <- get(x = paste0("M_DONOR_", k))
  bromo <- M_DONOR_ %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(QuantSN == max(QuantSN))
  bromo_Morgue[[k]] <- bromo 
}
names(bromo_Morgue) <- paste0("DONOR_", 1:9)
```


```{r}
bromo_REST <- vector("list", length = 9)
for(k in c(1:4, 6:9)) {
  load(paste0("./dat_un/RH", k, ".rda"))
  RH_ <- get(x = paste0("RH", k))
  bromo <- RH_ %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(QuantSN == max(QuantSN))
  bromo_REST[[k]] <- bromo
}
names(bromo_REST) <- paste0("RH_", 1:9)
```

## Cleanup, Make Excel

```{r}
s1 <- do.call("rbind", bromo_Morgue)
s1 <- s1 %>% select(Sample, RT1, RT2)
s2 <- do.call("rbind", bromo_REST)
s2 <- s2 %>% select(Sample, RT1, RT2)
bromo <- rbind(s1, s2)
bromo <- t(bromo)
bromo <- data.frame(bromo)
names(bromo) <- bromo[1, ]
bromo <- bromo[-1, ]
bromo <- apply(bromo, MAR = 2, FUN = as.numeric)
bromo <- as.data.frame(bromo)
bromo
```

## Write Out File

```{r}
write.csv(file = "Checking_Bromobenze_IS_un.csv", bromo, row.names = FALSE)
```


```{r}
plot(x = unlist(bromo[1, ]),
     y = unlist(bromo[2, ]),
     xlab = "RT1", 
     ylab = "RT2")
```

## Extract, Find Medians in Both Dimensions - from Aligned

```{r}
bromo_Morgue <- vector("list", length = 9)
for(k in c(1:4, 6:9)) {
  load(paste0("./dat/M_DONOR_", k, ".rda"))
  M_DONOR_ <- get(x = paste0("M_DONOR_", k))
  bromo <- M_DONOR_ %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           top_n(wt = PeakSN, n = 2) %>%
           group_split()
  bromo_Morgue[[k]] <- do.call("rbind", bromo)
}
names(bromo_Morgue) <- paste0("DONOR_", 1:9)

bromo_REST <- vector("list", length = 9)
for(k in c(1:4, 6:9)) {
  load(paste0("./dat/RH", k, ".rda"))
  RH_ <- get(x = paste0("RH", k))
  bromo <- RH_ %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           top_n(wt = PeakSN, n = 2) %>%
           group_split()
  bromo_REST[[k]] <- do.call("rbind", bromo)
}
names(bromo_REST) <- paste0("RH_", 1:9)

bromo <- rbind(do.call("rbind", bromo_Morgue) %>% select(RT1, RT2, BaseMass, PeakSN), 
               do.call("rbind", bromo_REST) %>% select(RT1, RT2, BaseMass, PeakSN))
```

From this, we find that we have `r nrow(bromo)` recorded records of Bromobenzene
in our data set. The median RTs are `r median(bromo$RT1)` and `r median(bromo$RT2)`.
Plotted, they look like:

```{r}
cols <- rep("black", nrow(bromo))
cols[bromo$BaseMass <= 77.10 & bromo$BaseMass >= 76.90] <- "red"
plot(bromo$RT1, bromo$RT2, col = cols)
```

```{r}
#
#  Function which runs a sliding rectangular window over a plot of RT1/RT2,
#  and finds the window with the highest number of incidences of common-mass
#  compound inside it. This is then centroided, and used as the RT centroid
#  for isolation of compounds in the following steps.
#
sliding_window <- function(dat, xww, yww) {

  # dat has RT1, RT2 and BaseMass in it - find the most common mass
  filt <- hist(dat$BaseMass, breaks = 1:(max(dat$BaseMass) + 1),
               plot = FALSE)
  common_mass <- filt$mids[filt$counts == max(filt$counts)][1]
  dat <- dat %>% filter(BaseMass <= common_mass * 1.02, 
                        BaseMass >= common_mass * 0.98)
  
  min_x <- min(dat$RT1)
  max_x <- max(dat$RT1)
  min_y <- min(dat$RT2)
  max_y <- max(dat$RT2)
  
  # Form a set of bottom-left to top-right boundaries, coarse - half the 
  # widths xww and yww as steps
  x_seq <- seq(from = min_x - xww, to = max_x + xww, by = xww / 2)
  y_seq <- seq(from = min_y - yww, to = max_y + yww, length.out = length(x_seq))
  boundaries1 <- data.frame(x1 = rep(x_seq, each = length(y_seq)), 
                            x2 = rep(x_seq + xww, each = length(y_seq)),
                            y1 = rep(y_seq, length(x_seq)),
                            y2 = rep(y_seq + yww, length(x_seq)))
  first_pass <- apply(boundaries1, MAR = 1, FUN = function(z) {
                        nrow(dat[dat$RT1 >= z[1] & dat$RT1 <= z[2] & 
                             dat$RT2 >= z[3] & dat$RT2 <= z[4], ])
                    })
  
  # identify the area of local maximum, then zoom in there and redo this
  # with a much finer x-resolution (that's the limiting factor)
  which_max <- which(first_pass == max(first_pass))[1]
  x_seq <- seq(from = boundaries1[which_max, "x1"] - 2 * xww,
               to = boundaries1[which_max, "x2"] + xww,
               by = xww / 10)
  y_seq <- seq(from = boundaries1[which_max, "y1"] - 2 * yww,
               to = boundaries1[which_max, "y2"] + yww,
               by = yww / 10)
  boundaries2 <- data.frame(x1 = rep(x_seq, each = length(y_seq)), 
                            x2 = rep(x_seq + xww, each = length(y_seq)),
                            y1 = rep(y_seq, length(x_seq)),
                            y2 = rep(y_seq + yww, length(x_seq)))
  
  second_pass <- apply(boundaries2, MAR = 1, FUN = function(z) {
                         nrow(dat[dat$RT1 >= z[1] & dat$RT1 <= z[2] & 
                              dat$RT2 >= z[3] & dat$RT2 <= z[4], ])
                     })
  x <- mean(unlist(boundaries2[which(second_pass == max(second_pass))[1], 1:2]))
  y <- mean(unlist(boundaries2[which(second_pass == max(second_pass))[1], 3:4]))
  return(c(common_mass, round(x, 2), round(y, 2)))
}
```

```{r testing}
sliding_window(bromo, 10, 0.05)
sliding_window(bromo, 10, 0.10)
sliding_window(bromo, 15, 0.05)
sliding_window(bromo, 15, 0.10)
```

This is the estimated 'centroid' for the common mass (+/-2%),
and the centroid RT1 and RT2 estimates, which match your chosen RT1 and RT2
variability.

We can now apply this method to any compound we like, to find its "definitive"
window, which can then be used for filtering! Time to write some more code ...

```{r}
extract_compound <- function(c_name) {
  Morgue <- vector("list", length = 9)
  for(k in c(1:4, 6:9)) {
    if(!exists(paste0("M_DONOR_", k))) {
      load(paste0("./dat/M_DONOR_", k, ".rda"))
    }
    M_DONOR_ <- get(x = paste0("M_DONOR_", k))
    temp <- M_DONOR_ %>% 
            subset(substr(Name, 1, nchar(c_name)) == c_name)
    Morgue[[k]] <- temp
  }

  REST <- vector("list", length = 9)
  for(k in c(1:4, 6:9)) {
    if(!exists(paste0("RH", k))) {
      load(paste0("./dat/RH", k, ".rda"))
    }
    RH_ <- get(x = paste0("RH", k))
    temp <- RH_ %>% 
            subset(substr(Name, 1, nchar(c_name)) == c_name) # remember the (E), (Z) stuff
    REST[[k]] <- temp
  }
  m <- do.call("rbind", Morgue) 
  if(!is.null(m)) {
    m <- m %>% select(RT1, RT2, BaseMass, PeakSN)
  } 
  r <- do.call("rbind", REST) 
  if(!is.null(r)) { 
    r <- r %>% select(RT1, RT2, BaseMass, PeakSN)
  }
  compound <- rbind(m, r)
  compound
}
```

## Applying Commonality Checks to All Compounds

We have a large series of compounds, all of which we don't know which should be
considered the true QuantMass, or the window of RTs where they should be considered.
We can use the above function to isolate the data-driven answer to this
for each compound, and then save it. This can then be used to repeat all of the
previous work, isolating on these findings **first**, and only then working through
the identification of "best peak per sample" and coincidences.

The compounds we are interested in are all located in pca_ready_all.csv
and Unaligned_pca_ready_all.csv. These were the outputs from Step 5, after 
our first pass. 

```{r}
compounds_align <- read.csv("pca_ready_all.csv")
compounds_unalign <- read.csv("Unaligned_pca_ready_all.csv")
compounds <- unique(c(compounds_align$Name, compounds_unalign$Name))
```

Iterate through these compounds, and for each, extract the list of compounds
across samples. This will be not filtered - simply every detected compound
of interest. The function sliding_window will isolate the most common
shared mass, filter out all non-matching identifications, and then identify
the centroid of the RT window that is most appropriate. These can then be
saved as three values together with the compound name. Another apply!

```{r}

# ref_comp <- apply(filter_results, MAR = 1, FUN = function(z) {
#                         dat <- extract_compound(z[1])
#                         c(z[1], sliding_window(dat, 10, 0.2))
#                  })

# same code, but easier to debug - was having issues with non-detects due to
# the (Z), (E) stuff - fixed above.
ref_comp <- data.frame(Name = compounds,
                       common_mass = NA,
                       window_RT1 = NA,
                       window_RT2 = NA)
for(j in 1:length(compounds)) {
  dat <- extract_compound(compounds[j])
  ref_comp[j, ] <- c(compounds[j], sliding_window(dat, 10, 0.2))
}
```

Now, save out this object - we will use it in the future. 

```{r}
save(file = "compound_ref_RT_mass.rda", ref_comp)
write.csv(x = ref_comp, file = "references_for_compounds.csv", 
          row.names = FALSE)
```
